Seperate logic from UI
